Changing the copy for the Homepage, moving some components to header folder
