+++ src/components/header/Header.js
@@ -4,7 +4,7 @@ import { FormattedMessage } from 'react-intl'
 import Container from '@material-ui/core/Container'
 import Typography from '@material-ui/core/Typography'
 import Grid from '@material-ui/core/Grid'
-import { useStyles } from '../style/HeaderStyle'
+import { useStyles } from '../../style/HeaderStyle'
 
 const Header = () => {
   const classes = useStyles()
@@ -22,7 +22,7 @@ const Header = () => {
               <Information position={'left'} />
             </Grid>
             <Grid item md={6}>
-              <Information position={'right'} />
+              <Information position={'right'} />              
             </Grid>
           </Grid>
         </div>
