+++ src/translations/en.json
@@ -6,11 +6,11 @@
   "information.left.deal": "* Deal directly with the end user. No middlemen and no fees.",
   "information.left.title": "Makers",
   "information.left.button": "BROWSE WHAT’S NEEDED",
-  "information.right.visualize": "* Visualize and locate real, updated hospital needs. Start helping today!",
-  "information.right.learn": "* Learn about specifications. No more rejections of unusable material.",
-  "information.right.deal": "* Deal directly with the end user. No middlemen and no fees.",
+  "information.right.visualize": "* Are you a hospital or nursing home manager interested in listing your PPE needs",
+  "information.right.learn": " on this site?",
+  "information.right.deal": "Contact us at mensageocov@gmail.com",
   "information.right.title": "Health Professionals",
-  "information.right.button": "LIST YOUR NEEDS",
+  "information.right.button": "CONTACT US",
   "header.title": "We connect health organizations with makers of critical equipment.",
   "need.update.title": "There’s a need for...",
   "need.update.last": "Last Updated: ",
