+++ src/translations/es.json
@@ -6,9 +6,9 @@
 	"information.left.deal": "* Trata directamente con el usuario final. Sin intermediarios ni comisiones.",
 	"information.left.title": "Makers",
 	"information.left.button": "NAVEGA POR LAS NECESIDADES",
-	"information.right.visualize": "* Visualiza y encuentra las necesidades actuales de los centros sanitarios. Comienza a ayudar hoy!",
-	"information.right.learn": "* Aprende sobre las especificaciones. No más rechazos por material no válido.",
-	"information.right.deal": "* Trata directamente con el usuario final. Sin intermediarios ni comisiones.",
+	"information.right.visualize": "* Are you a hospital or nursing home manager interested in listing your PPE needs",
+  	"information.right.learn": " on this site?",
+  	"information.right.deal": "Contact us at mensageocov@gmail.com",
 	"information.right.title": "Profesionales de la salud",
 	"information.right.button": "INDICA TUS NECESIDADES",
 	"header.title": "Conectamos organizaciones sanitarias con creadores de equipaciones críticas.",
